+++ frontend/src/hooks/domain/photos/useFileUpload.ts
