+++ frontend/src/hooks/domain/photos/usePhotosBySpaceCode.ts
