+++ frontend/src/hooks/domain/photos/usePhotosDelete.ts
