+++ frontend/src/hooks/domain/space/useSpaceCodeFromPath.ts
